requirements.txt
annotated-types==0.6.0
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.8.0
FLAML==2.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.2
openai==1.3.9
pyautogen==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
